Provide finer task state tracking to differentiate step start and end
Send and display tool call start the moment the model makes the call.
Update the train of thought with tool call result on completion.

This gives more granular visibility to user on which task is currently
being executed. It is especially useful to track long running tool
calls.

Previously we'd only show the tool call after it completed. So there
was no visibility into what was taking so long.

Now we can see the tool call that is taking time to complete and react
to that by either giving different guidance or pausing the task